refactor : -  moved ds to separate package - remove todo and comment codes
