--- com/jk/game/DiagonalSum.java
@@ -1,5 +1,0 @@
-package com.jk.game;
-
-public class DiagonalSum {
-    int sumOfD1, sumOfD2;
-}
