+++ com/jk/game/Game.java
@@ -1,5 +1,7 @@
 package com.jk.game;
 
+import com.jk.game.ds.DiagonalSum;
+import com.jk.game.ds.Position;
 
 class Game {
