+++ com/jk/game/ds/DiagonalSum.java
@@ -1,0 +1,6 @@
+package com.jk.game.ds;
+
+public class DiagonalSum {
+    public int sumOfD1;
+	public int sumOfD2;
+}
